add a sleep to ram_write
Sometimes ram_write returns a NAK, so we need to look at the return code
before moving onto the next command. But if we don't sleep, then we
sometimes (not often) get an empty message before we get the ACK/NAK. We
could retry, but I've hit a case where part of the real message is
transmitted after the empty message during the same transaction, and
since the monitor doesn't buffer its responses, the real message is
lost. So the only thing we can do is pay the penalty of the sleep for
every message. Sadness :(.