add command-and-control monitor payload
This programs the IMP pixel grab hardware, hooks the vsync interrupt to
grab the captured pixel value and append it to a buffer, and hooks the
main loop in order to process the resulting packet that's been received
through 3 blinking pixels on the screen. We use 3 pixels instead of just
one since we often don't have exact control over framerate, and
decreasing the maximum framerate by 3 gives us more slack to make sure
we don't go over it without slowing things down too much.

The packets are currently sent to the monitor through a single gif that
can be embedded in web pages. The gif has to be at a hardcoded offset
specified by cnc.s. cnc_packet.py has code for generating the gif, as
well as crafting the packets that go inside it.