delltools: add the ability to show multiple images at once
In order to show two images at once, there are a few things we need to
do:
1. Put the images in different tiles.
2. Put the actual image data in a different sdram location.
3. Offset the second image's indices so that the images can share the
same CLUT.

This commit modifies put_image so that it can do #1 and #2. #3 has to be
done both at image upload time (modify the actual data) and at put_image
time (push the CLUT data at an offset), so the CLUT offset is a
parameter of upload_single_image. upload_all_images currently always
uses a CLUT offset of 0, since it isn't meant for displaying multiple
images at the same time.